gemini/05-chatbots/5.1 gemini with gradio/gradio-svc/requirements.txt
gradio
Pillow
google-cloud-aiplatform